move new tests into dedicated unit test folder
